Revert "[fix] fix cargo fmt invocation in CI script"
This reverts commit 041ef7e284239b6f49a820df63f31fb6d4268ff1.

Didn't realize you needed the "--" to pass arguments from cargo fmt
to rustfmt.
It's not the invocation of rustfmt that's broken, it seems that
rustfmt on the CI server is using a different set of settings or
something.

My local checkout of the code does cargo rustfmt -- --write-mode=diff
and reports no changes, but the CI runs the same command and finds
things to change.
hmmmmmm