[refactor] genericize device read/write functions in prep for other modules
Previously the {read, write}_file functions were basically tailored to
the GPIO module.
With PWM waiting in the wings, those need to be genericized so that
the PWM module can call them.
If we didn't do this, we'd have to write PWM-specific functions, which
is just a waste of time.

Still looking for a way to make this a little more elegant though.